cambios de prueba

+++ shop-microservice/src/main/java/com/integrador5/shopmicroservice/DTO/MostPopularProductDTO.java
@@ -3,9 +3,9 @@ package com.integrador5.shopmicroservice.DTO;
 public class MostPopularProductDTO {
     private int idProduct;
     private String nameProduct;
-    private long quantity;
+    private int quantity;
 
-    public MostPopularProductDTO(int idProduct, String nameProduct, long quantity) {
+    public MostPopularProductDTO(int idProduct, String nameProduct, int quantity) {
         this.idProduct = idProduct;
         this.nameProduct = nameProduct;
         this.quantity = quantity;
